pre-merge with updated DEV (Front Log connected)
