+++ src/components/FrontLog.tsx
@@ -48,21 +48,21 @@ const requests = [
   // More requests...
 ];
 
-const socket = new WebSocket('ws://localhost:1234');
+// const socket = new WebSocket('ws://localhost:1234');
 
-socket.addEventListener('open', (event) => {
-  console.log('CONNECTED TO WEB SOCKET FROM CLIENT Side');
-});
+// socket.addEventListener('open', (event) => {
+//   console.log('CONNECTED TO WEB SOCKET FROM CLIENT Side');
+// });
 
-socket.addEventListener('message', (event) => {
-  // logic to display received data here
-  // likely use state components
-  console.log('MESSAGE RECEIVED FROM 1234: ', event.data);
-});
+// socket.addEventListener('message', (event) => {
+//   // logic to display received data here
+//   // likely use state components
+//   console.log('MESSAGE RECEIVED FROM 1234: ', event.data);
+// });
 
-const sendMessage = () => {
-  socket.send('1. CLIENT 1 JUST SEND THIS MESSAGE TO SERVER!!!!!!');
-};
+// const sendMessage = () => {
+//   socket.send('1. CLIENT 1 JUST SEND THIS MESSAGE TO SERVER!!!!!!');
+// };
 
 export default function FrontLog() {
   return (
@@ -74,9 +74,9 @@ export default function FrontLog() {
               Request Log
             </h3>
             <div>
-              <button onClick={sendMessage} style={{ color: 'white' }}>
+              {/* <button onClick={sendMessage} style={{ color: 'white' }}>
                 SEND MESSAGE
-              </button>
+              </button> */}
             </div>
           </div>
         </div>
