+++ src/containers/FrontTester.tsx
@@ -1,41 +1,33 @@
 import FrontLog from '../components/FrontLog';
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
 
-// const socket = new WebSocket('ws://localhost:6969');
-
-// socket.addEventListener('open', (event) => {
-//   console.log('Client 2: CONNECTED TO WEB SOCKET FROM CLIENT Side');
-// });
-
-// socket.addEventListener('message', (event) => {
-//   console.log('Client 2: MESSAGE RECEIVED FROM 6969: ', event.data);
-// });
+const socket = new WebSocket('ws://localhost:1234');
 
 export default function FrontTester() {
-  // const client = new WebSocket('ws://localhost:1234');
+  const [message, updateMessage] = useState([]);
+  socket.addEventListener('open', (event) => {
+    console.log('CONNECTED TO WEB SOCKET FROM CLIENT Side');
+  });
 
-  // client.addEventListener('open', function (event) {
-  //   client.send('Hello Backend!');
-  // });
+  socket.addEventListener('message', (event) => {
+    // logic to display received data here
+    // likely use state components
+    console.log('MESSAGE RECEIVED FROM 1234: ', event.data);
+    updateMessage([...message, event.data]);
+  });
 
-  // client.addEventListener('message', function (event) {
-  //   console.log('Message from ze back', event.data);
-  // });
-
-  useEffect(() => {
-    // const client = new WebSocket('ws://localhost1234')
-    // client.onopen = () => {
-    //   console.log('Connected to socket');
-    // }
-    // client.onmessage = (event) => {
-    //   const message = event.data;
-    //   console.log(message);
-    // }
-  }, []);
+  const sendMessage = () => {
+    socket.send('1. CLIENT 1 JUST SEND THIS MESSAGE TO SERVER!!!!');
+  };
+  useEffect(() => {}, []);
 
   return (
     <div className='bg-gray-900 h-screen'>
       <FrontLog />
+      <button onClick={sendMessage} style={{ color: 'white' }}>
+        SEND MESSAGE
+      </button>
+      <div style={{ color: 'white' }}>{message}</div>
     </div>
     // <div className="front-tester-container">
     //   <div className="api-dropdown"></div>
